# internal/handlers/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Relational Knowledge Engineering Platform API
  description: |
    A comprehensive API for document processing, knowledge extraction, and graph-based analysis.
    
    The platform supports:
    - PDF document upload and processing
    - Chunked file uploads for large documents
    - Text extraction and NLP processing
    - Knowledge graph generation and analysis
    - Document management and retrieval
    
    ## Authentication
    Currently, the API does not require authentication for public endpoints.
    
    ## Rate Limiting
    No rate limiting is currently implemented.
    
    ## Error Handling
    All endpoints return consistent error responses with appropriate HTTP status codes.
  version: 1.0.0
  contact:
    name: Relational Knowledge Engineering Platform
    url: https://github.com/yothgewalt/relational-knowledge-engineering-platform-server
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Development server
  - url: https://api.example.com
    description: Production server
...
paths:
...
  /api/v1/upload/initiate:
    post:
      tags: [Upload]
      summary: Initiate chunked upload
      description: |
        Initiates a new chunked upload session for large files.
        Returns session information including chunk size and upload ID.
      operationId: initiateUpload
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InitiateUploadRequest'
      responses:
        '200':
          description: Upload session initiated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  session_id:
                    type: string
                    format: uuid
                    example: "123e4567-e89b-12d3-a456-426614174000"
                  upload_id:
                    type: string
                    example: "upload-123456"
                  chunk_size:
                    type: integer
                    format: int64
                    example: 5242880
                  total_chunks:
                    type: integer
                    example: 10
                  minio_key:
                    type: string
                    example: "uploads/123e4567-e89b-12d3-a456-426614174000/document.pdf"
                  message:
                    type: string
                    example: "Upload session initiated successfully"
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/upload/chunk:
    post:
      tags: [Upload]
      summary: Upload file chunk
      description: |
        Uploads a single chunk of a file to an existing upload session.
        Chunks must be uploaded in order.
      operationId: uploadChunk
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                session_id:
                  type: string
                  format: uuid
                  description: Upload session ID
                chunk_index:
                  type: integer
                  description: Index of the chunk (starting from 1)
                chunk:
                  type: string
                  format: binary
                  description: Binary chunk data
              required:
                - session_id
                - chunk_index
                - chunk
      responses:
        '200':
          description: Chunk uploaded successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  session_id:
                    type: string
                    format: uuid
                  chunk_index:
                    type: integer
                  chunk_etag:
                    type: string
                  uploaded_chunks:
                    type: integer
                  total_chunks:
                    type: integer
                  message:
                    type: string
                    example: "Chunk uploaded successfully"
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/upload/{sessionId}/complete:
    post:
      tags: [Upload]
      summary: Complete chunked upload
      description: |
        Completes a chunked upload session by combining all uploaded chunks.
        All chunks must be uploaded before calling this endpoint.
      operationId: completeUpload
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Upload session ID
      responses:
        '200':
          description: Upload completed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  session_id:
                    type: string
                    format: uuid
                  file_info:
                    $ref: '#/components/schemas/FileInfo'
                  message:
                    type: string
                    example: "Upload completed successfully"
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/upload/{sessionId}/abort:
    delete:
      tags: [Upload]
      summary: Abort chunked upload
      description: |
        Aborts an active upload session and cleans up any uploaded chunks.
        This operation cannot be undone.
      operationId: abortUpload
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Upload session ID
      responses:
        '200':
          description: Upload aborted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  session_id:
                    type: string
                    format: uuid
                  message:
                    type: string
                    example: "Upload aborted successfully"
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/upload/{sessionId}/progress:
    get:
      tags: [Upload]
      summary: Get upload progress
      description: |
        Returns the current progress of an upload session including
        uploaded chunks, total progress, and session status.
      operationId: getUploadProgress
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Upload session ID
      responses:
        '200':
          description: Upload progress retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadProgress'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  schemas:
    InitiateUploadRequest:
      type: object
      required:
        - filename
        - file_size
        - content_type
      properties:
        filename:
          type: string
          example: "research-paper.pdf"
          description: Name of the file to upload
        file_size:
          type: integer
          format: int64
          example: 52428800
          description: Size of the file in bytes
        content_type:
          type: string
          example: "application/pdf"
          description: MIME type of the file
        chunk_size:
          type: integer
          format: int64
          example: 5242880
          description: Preferred chunk size in bytes (optional)
        metadata:
          type: object
          additionalProperties:
            type: string
          example:
            author: "John Doe"
            category: "research"
          description: Additional metadata for the file
    UploadProgress:
      type: object
      properties:
        session_id:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        filename:
          type: string
          example: "research-paper.pdf"
        file_size:
          type: integer
          format: int64
          example: 52428800
        uploaded_bytes:
          type: integer
          format: int64
          example: 26214400
        completed_chunks:
          type: integer
          example: 5
        total_chunks:
          type: integer
          example: 10
        progress:
          type: number
          format: float
          minimum: 0
          maximum: 100
          example: 50.0
        status:
          type: string
          enum: [active, completed, failed, aborted]
          example: "active"
        created_at:
          type: string
          format: date-time
          example: "2023-12-01T10:00:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2023-12-01T10:30:00Z"
...
    FileInfo:
      type: object
      properties:
        key:
          type: string
          example: "uploads/123e4567-e89b-12d3-a456-426614174000/document.pdf"
        size:
          type: integer
          format: int64
          example: 52428800
        content_type:
          type: string
          example: "application/pdf"
        last_modified:
          type: string
          format: date-time
          example: "2023-12-01T10:00:00Z"
        etag:
          type: string
          example: "d41d8cd98f00b204e9800998ecf8427e"
        metadata:
          type: object
          additionalProperties:
            type: string
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
        code:
          type: integer
          description: HTTP status code
        details:
          type: object
          description: Additional error details
          nullable: true
  responses:
    BadRequest:
      description: Bad request - invalid input parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Invalid request parameters"
            code: 400
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Resource not found"
            code: 404
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Internal server error"
            code: 500
...